Move and improve connection tests
